+++ skeleton-saints-ahi/Assets/Scripts/IDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamage
 {
-    void TakeDamage(int damage);
+    void TakeDamage(float damage);
 }
